+++ playground/wizard/src/components/NavBar.tsx
@@ -1,50 +1,49 @@
-import { Dispatch, SetStateAction } from "react";
-import { useDarkMode } from "../hooks/useDarkMode";
-import { Toggle } from "pentatrion-design/components/input/Toggle";
-import { Select } from "pentatrion-design/components/select";
-import LinkButton from "./LinkButton";
+import clsx from "clsx";
+
+import { ResizeArea } from "pentatrion-design/components/resize-area";
+import { Button, LinkButton } from "pentatrion-design/components/button";
+
+import { Dispatch, ReactNode, SetStateAction } from "react";
 
 interface Props {
-  theme: ThemeID;
-  setTheme: Dispatch<SetStateAction<ThemeID>>;
+  showNavBar: boolean;
+  setShowNavBar: Dispatch<SetStateAction<boolean>>;
+  children: ReactNode;
 }
-
-const themeOptions = [
-  { value: "legacy", label: "Current" },
-  { value: "classic", label: "Classic" },
-  { value: "modern", label: "Modern" },
-];
-
-export default function NavBar({ theme, setTheme }: Props) {
-  const { isDarkMode, setDarkMode } = useDarkMode();
+export default function NavBar({ showNavBar, setShowNavBar, children }: Props) {
   return (
-    <div className="flex-1 flex flex-col">
-      <h1 className="h-8 flex items-center mb-4">
-        <span className="app-logo text-2xl">
-          Map<span className="text-[#82b4fe]">Libre</span>
-        </span>
-      </h1>
-      <div className="flex flex-col gap-2 flex-1">
-        <div className="p8n-setting">
-          <div>Dark mode</div>
-          <Toggle
-            checked={isDarkMode}
-            onChange={(e) => setDarkMode(e.target.checked)}
-          />
+    <div
+      id="menu-col"
+      className={clsx(
+        "fixed left-0 top-0 z-30 h-screen flex-col max-w-full w-64 flex-none  bg-gray-0  shadow dark:shadow-dark md:relative md:flex md:bg-transparent md:shadow-none md:dark:shadow-none md:flex-[0_0_var(--sidebar-menu-width)] border-r border-r-gray-2",
+        !showNavBar ? "hidden" : "flex"
+      )}
+    >
+      <ResizeArea name="menu" position="right" className="hidden md:block" />
+      <header className="h-8 flex items-center m-2 justify-between">
+        <a href="https://maplibre.org" className="block w-full">
+          <h1 className="flex items-center justify-between">
+            <span className=" text-2xl app-logo">
+              Map<span className="text-[#82b4fe]">Libre</span>
+            </span>
+            <span className="text-gray-6">
+              <i className="fe-sliders"></i>
+              Theme customizer
+            </span>
+          </h1>
+        </a>
+        <div className="md:hidden">
+          <Button
+            icon
+            onClick={() => setShowNavBar(false)}
+            color="gray"
+            variant="text"
+          >
+            <i className="fe-cancel"></i>
+          </Button>
         </div>
-        <div className="p8n-setting">
-          <div>Theme</div>
-          <Select
-            variant="ghost"
-            options={themeOptions}
-            value={theme}
-            onChange={(o) => {
-              // @ts-expect-error value is ThemeID
-              setTheme(o.target.value);
-            }}
-          ></Select>
-        </div>
-      </div>
+      </header>
+      {children}
       <div className="flex items-center justify-center h-12">
         <LinkButton
           icon