+++ playground/wizard/src/lib/css.ts
@@ -1,0 +1,18 @@
+export function mergeCssValues(
+  cssValues: CssValues | null,
+  isDarkMode: boolean
+): {
+  [key: string]: string;
+} {
+  console.log("mergeCssValues");
+  if (!cssValues) {
+    return {};
+  }
+  if (isDarkMode) {
+    return {
+      ...cssValues["light"],
+      ...cssValues["dark"],
+    };
+  }
+  return cssValues["light"];
+}