+++ packages/maplibre-theme/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.8
+
+- Modern/Classic : rename CSS color name `gray` to `primary`.
+- Modern : fix `--ml-ctrl-border-radius`
+- Classic : fix popup shadow
+
 ## 0.0.7
 
 - Modern : remove rounded map container